+++ src/Former/Traits/Field.php
@@ -7,13 +7,14 @@
  */
 namespace Former\Traits;
 
-use \Former\Form\Form;
-use \Former\Form\Group;
-use \Former\Helpers;
-use \Former\Interfaces\FieldInterface;
-use \Former\LiveValidation;
-use \Underscore\Types\Arrays;
-use \Underscore\Types\String;
+use Former\Former;
+use Former\Form\Form;
+use Former\Form\Group;
+use Former\Helpers;
+use Former\Interfaces\FieldInterface;
+use Former\LiveValidation;
+use Underscore\Types\Arrays;
+use Underscore\Types\String;
 
 abstract class Field extends FormerObject implements FieldInterface
 {
@@ -67,10 +68,15 @@ abstract class Field extends FormerObject implements FieldInterface
     $this->type       = $type;
     $this->value      = $value;
 
-    // Set magic parameters (repopulated value, translated label, etc)
+    // Compute and translate label
     $this->automaticLabels($name, $label);
 
-    if($type != 'password') $this->value = $this->repopulate();
+    // Repopulate field
+    if($type != 'password') {
+      $this->value = $this->repopulate();
+    }
+
+    // Apply Live validation rules
     if ($this->app['former']->getOption('live_validation')) {
       $rules = new LiveValidation($this);
       $rules->apply($this->getRules());
@@ -78,7 +84,10 @@ abstract class Field extends FormerObject implements FieldInterface
 
     // Link Control group
     if ($this->app['former']->getFramework()->isnt('Nude')) {
-      $this->group = new Group($this->app, $this->label);
+      $groupClass = $this->isCheckable() ? 'CheckableGroup' : 'Group';
+      $groupClass = Former::FORMSPACE.$groupClass;
+
+      $this->group = new $groupClass($this->app, $this->label);
     }
   }
 
@@ -122,7 +131,7 @@ abstract class Field extends FormerObject implements FieldInterface
   }
 
   ////////////////////////////////////////////////////////////////////
-  ///////////////////////////// FUNCTIONS ////////////////////////////
+  ////////////////////////// PUBLIC INTERFACE ////////////////////////
   ////////////////////////////////////////////////////////////////////
 
   /**
@@ -286,10 +295,10 @@ abstract class Field extends FormerObject implements FieldInterface
     $populate = $this->app['former']->getValue($this->name);
 
     // Assign a priority to each
-    if(!is_null($post)) $value = $post;
-    elseif(!is_null($populate)) $value = $populate;
-    else $value = $fallback;
-    return $value;
+    if(!is_null($post)) return $post;
+    if(!is_null($populate)) return $populate;
+
+    return $fallback;
   }
 
   /**