+++ tests/_start.php
@@ -1,5 +1,8 @@
 <?php
+ini_set('memory_limit', '120M');
 include '_illuminate.php';
+include 'Dummy/DummyButton.php';
+include 'Dummy/DummyEloquent.php';
 
 // Base Test class for matchers
 abstract class FormerTests extends PHPUnit_Framework_TestCase